move project functions into Projects.js, create displayProjects()
