+++ src/index.js
@@ -1,10 +1,7 @@
 import "./styles.css";
 
 import displayTodos from "./components/TodoItems";
-import Projects from "./components/Projects";
-
-//test data to delete later
-import { projectsArray } from "./testData";
+import displayProjects from "./components/Projects";
 
 function runApp() {
   displayContent();
@@ -15,10 +12,5 @@ runApp();
 // prob separate this into its own util function / module later
 function displayContent() {
   displayTodos();
-  //displays projects - just pass in all the arrays
-  const projectsListContainer = document.querySelector(
-    ".projects-list-container"
-  );
-  const projectsList = Projects(projectsArray);
-  projectsListContainer.appendChild(projectsList);
+  displayProjects();
 }
