--- src/testData.js
@@ -1,12 +1,0 @@
-export const projectsArray = [
-  {
-    id: 0,
-    name: "inbox",
-    active: true,
-  },
-  {
-    id: 1,
-    name: "cooking",
-    active: false,
-  },
-];
